+++ barretenberg/cpp/src/barretenberg/honk/proof_system/ultra_prover.cpp
@@ -1,9 +1,6 @@
 #include "ultra_prover.hpp"
-#include "barretenberg/honk/pcs/claim.hpp"
 #include "barretenberg/honk/sumcheck/sumcheck.hpp"
 #include "barretenberg/honk/utils/power_polynomial.hpp"
-#include "barretenberg/polynomials/polynomial.hpp"
-#include "barretenberg/transcript/transcript_wrappers.hpp"
 
 namespace proof_system::honk {
 
@@ -16,9 +13,8 @@ namespace proof_system::honk {
  * */
 template <UltraFlavor Flavor>
 UltraProver_<Flavor>::UltraProver_(std::shared_ptr<Instance> inst)
-    : queue(inst->commitment_key, transcript)
-    , instance(std::move(inst))
-    , pcs_commitment_key(instance->commitment_key)
+    : instance(std::move(inst))
+    , commitment_key(instance->commitment_key)
 {
     instance->initialise_prover_polynomials();
 }
@@ -54,14 +50,14 @@ template <UltraFlavor Flavor> void UltraProver_<Flavor>::execute_wire_commitment
     auto wire_polys = instance->proving_key->get_wires();
     auto labels = commitment_labels.get_wires();
     for (size_t idx = 0; idx < 3; ++idx) {
-        queue.add_commitment(wire_polys[idx], labels[idx]);
+        transcript.send_to_verifier(labels[idx], commitment_key->commit(wire_polys[idx]));
     }
 
     if constexpr (IsGoblinFlavor<Flavor>) {
         auto op_wire_polys = instance->proving_key->get_ecc_op_wires();
         auto labels = commitment_labels.get_ecc_op_wires();
         for (size_t idx = 0; idx < Flavor::NUM_WIRES; ++idx) {
-            queue.add_commitment(op_wire_polys[idx], labels[idx]);
+            transcript.send_to_verifier(labels[idx], commitment_key->commit(op_wire_polys[idx]));
         }
     }
 }
@@ -78,8 +74,10 @@ template <UltraFlavor Flavor> void UltraProver_<Flavor>::execute_sorted_list_acc
 
     // Commit to the sorted withness-table accumulator and the finalised (i.e. with memory records) fourth wire
     // polynomial
-    queue.add_commitment(instance->proving_key->sorted_accum, commitment_labels.sorted_accum);
-    queue.add_commitment(instance->proving_key->w_4, commitment_labels.w_4);
+    auto sorted_accum_commitment = commitment_key->commit(instance->proving_key->sorted_accum);
+    auto w_4_commitment = commitment_key->commit(instance->proving_key->w_4);
+    transcript.send_to_verifier(commitment_labels.sorted_accum, sorted_accum_commitment);
+    transcript.send_to_verifier(commitment_labels.w_4, w_4_commitment);
 }
 
 /**
@@ -93,8 +91,10 @@ template <UltraFlavor Flavor> void UltraProver_<Flavor>::execute_grand_product_c
 
     instance->compute_grand_product_polynomials(beta, gamma);
 
-    queue.add_commitment(instance->proving_key->z_perm, commitment_labels.z_perm);
-    queue.add_commitment(instance->proving_key->z_lookup, commitment_labels.z_lookup);
+    auto z_perm_commitment = commitment_key->commit(instance->proving_key->z_perm);
+    auto z_lookup_commitment = commitment_key->commit(instance->proving_key->z_lookup);
+    transcript.send_to_verifier(commitment_labels.z_perm, z_perm_commitment);
+    transcript.send_to_verifier(commitment_labels.z_lookup, z_lookup_commitment);
 }
 
 /**
@@ -111,99 +111,18 @@ template <UltraFlavor Flavor> void UltraProver_<Flavor>::execute_relation_check_
 }
 
 /**
- * - Get rho challenge
- * - Compute d+1 Fold polynomials and their evaluations.
+ * @brief Execute the ZeroMorph protocol to prove the multilinear evaluations produced by Sumcheck
+ * @details See https://hackmd.io/dlf9xEwhTQyE3hiGbq4FsA?view for a complete description of the unrolled protocol.
  *
  * */
-template <UltraFlavor Flavor> void UltraProver_<Flavor>::execute_univariatization_round()
+template <UltraFlavor Flavor> void UltraProver_<Flavor>::execute_zeromorph_rounds()
 {
-    const size_t NUM_POLYNOMIALS = Flavor::NUM_ALL_ENTITIES;
-
-    // Generate batching challenge ρ and powers 1,ρ,…,ρᵐ⁻¹
-    FF rho = transcript.get_challenge("rho");
-    std::vector<FF> rhos = pcs::gemini::powers_of_rho(rho, NUM_POLYNOMIALS);
-
-    // Batch the unshifted polynomials and the to-be-shifted polynomials using ρ
-    Polynomial batched_poly_unshifted(instance->proving_key->circuit_size); // batched unshifted polynomials
-    size_t poly_idx = 0;                                                    // TODO(#391) zip
-    for (auto& unshifted_poly : instance->prover_polynomials.get_unshifted()) {
-        batched_poly_unshifted.add_scaled(unshifted_poly, rhos[poly_idx]);
-        ++poly_idx;
-    }
-
-    Polynomial batched_poly_to_be_shifted(instance->proving_key->circuit_size); // batched to-be-shifted polynomials
-    for (auto& to_be_shifted_poly : instance->prover_polynomials.get_to_be_shifted()) {
-        batched_poly_to_be_shifted.add_scaled(to_be_shifted_poly, rhos[poly_idx]);
-        ++poly_idx;
-    };
-
-    // Compute d-1 polynomials Fold^(i), i = 1, ..., d-1.
-    gemini_polynomials = Gemini::compute_gemini_polynomials(
-        sumcheck_output.challenge, std::move(batched_poly_unshifted), std::move(batched_poly_to_be_shifted));
-
-    // Compute and add to trasnscript the commitments [Fold^(i)], i = 1, ..., d-1
-    for (size_t l = 0; l < instance->proving_key->log_circuit_size - 1; ++l) {
-        queue.add_commitment(gemini_polynomials[l + 2], "Gemini:FOLD_" + std::to_string(l + 1));
-    }
-}
-
-/**
- * - Do Fiat-Shamir to get "r" challenge
- * - Compute remaining two partially evaluated Fold polynomials Fold_{r}^(0) and Fold_{-r}^(0).
- * - Compute and aggregate opening pairs (challenge, evaluation) for each of d Fold polynomials.
- * - Add d-many Fold evaluations a_i, i = 0, ..., d-1 to the transcript, excluding eval of Fold_{r}^(0)
- * */
-template <UltraFlavor Flavor> void UltraProver_<Flavor>::execute_pcs_evaluation_round()
-{
-    const FF r_challenge = transcript.get_challenge("Gemini:r");
-    univariate_openings = Gemini::compute_fold_polynomial_evaluations(
-        sumcheck_output.challenge, std::move(gemini_polynomials), r_challenge);
-
-    for (size_t l = 0; l < instance->proving_key->log_circuit_size; ++l) {
-        std::string label = "Gemini:a_" + std::to_string(l);
-        const auto& evaluation = univariate_openings.opening_pairs[l + 1].evaluation;
-        transcript.send_to_verifier(label, evaluation);
-    }
-}
-
-/**
- * - Do Fiat-Shamir to get "nu" challenge.
- * - Compute commitment [Q]_1
- * */
-template <UltraFlavor Flavor> void UltraProver_<Flavor>::execute_shplonk_batched_quotient_round()
-{
-    nu_challenge = transcript.get_challenge("Shplonk:nu");
-
-    batched_quotient_Q = Shplonk::compute_batched_quotient(
-        univariate_openings.opening_pairs, univariate_openings.witnesses, nu_challenge);
-
-    // commit to Q(X) and add [Q] to the transcript
-    queue.add_commitment(batched_quotient_Q, "Shplonk:Q");
-}
-
-/**
- * - Do Fiat-Shamir to get "z" challenge.
- * - Compute polynomial Q(X) - Q_z(X)
- * */
-template <UltraFlavor Flavor> void UltraProver_<Flavor>::execute_shplonk_partial_evaluation_round()
-{
-    const FF z_challenge = transcript.get_challenge("Shplonk:z");
-
-    shplonk_output = Shplonk::compute_partially_evaluated_batched_quotient(univariate_openings.opening_pairs,
-                                                                           univariate_openings.witnesses,
-                                                                           std::move(batched_quotient_Q),
-                                                                           nu_challenge,
-                                                                           z_challenge);
-}
-/**
- * - Compute final PCS opening proof:
- * - For KZG, this is the quotient commitmecnt [W]_1
- * - For IPA, the vectors L and R
- * */
-template <UltraFlavor Flavor> void UltraProver_<Flavor>::execute_final_pcs_round()
-{
-    PCS::compute_opening_proof(pcs_commitment_key, shplonk_output.opening_pair, shplonk_output.witness, transcript);
-    // queue.add_commitment(quotient_W, "KZG:W");
+    ZeroMorph::prove(instance->prover_polynomials.get_unshifted(),
+                     instance->prover_polynomials.get_to_be_shifted(),
+                     sumcheck_output.claimed_evaluations,
+                     sumcheck_output.challenge,
+                     commitment_key,
+                     transcript);
 }
 
 template <UltraFlavor Flavor> plonk::proof& UltraProver_<Flavor>::export_proof()
@@ -219,48 +138,26 @@ template <UltraFlavor Flavor> plonk::proof& UltraProver_<Flavor>::construct_proo
 
     // Compute first three wire commitments
     execute_wire_commitments_round();
-    queue.process_queue();
 
     // Compute sorted list accumulator and commitment
     execute_sorted_list_accumulator_round();
-    queue.process_queue();
 
     // Fiat-Shamir: beta & gamma
     // Compute grand product(s) and commitments.
     execute_grand_product_computation_round();
-    queue.process_queue();
 
     // Fiat-Shamir: alpha
     // Run sumcheck subprotocol.
     execute_relation_check_rounds();
 
-    // Fiat-Shamir: rho
-    // Compute Fold polynomials and their commitments.
-    execute_univariatization_round();
-    queue.process_queue();
-
-    // Fiat-Shamir: r
-    // Compute Fold evaluations
-    execute_pcs_evaluation_round();
-
-    // Fiat-Shamir: nu
-    // Compute Shplonk batched quotient commitment Q
-    execute_shplonk_batched_quotient_round();
-    queue.process_queue();
-
-    // Fiat-Shamir: z
-    // Compute partial evaluation Q_z
-    execute_shplonk_partial_evaluation_round();
-
-    // Fiat-Shamir: z
-    // Compute PCS opening proof (either KZG quotient commitment or IPA opening proof)
-    execute_final_pcs_round();
+    // Fiat-Shamir: rho, y, x, z
+    // Execute Zeromorph multilinear PCS
+    execute_zeromorph_rounds();
 
     return export_proof();
 }
 
 template class UltraProver_<honk::flavor::Ultra>;
-template class UltraProver_<honk::flavor::UltraGrumpkin>;
 template class UltraProver_<honk::flavor::GoblinUltra>;
 
 } // namespace proof_system::honk