+++ yarn-project/world-state/src/synchronizer/server_world_state_synchronizer.ts
@@ -2,6 +2,7 @@ import { SerialQueue } from '@aztec/foundation/fifo';
 import { createDebugLogger } from '@aztec/foundation/log';
 import { elapsed } from '@aztec/foundation/timer';
 import { L2Block, L2BlockDownloader, L2BlockSource } from '@aztec/types';
+import { L2BlockHandledStats } from '@aztec/types/stats';
 
 import { LevelUp } from 'levelup';
 
@@ -172,7 +173,7 @@ export class ServerWorldStateSynchronizer implements WorldStateSynchronizer {
    */
   private async collectAndProcessBlocks() {
     // This request for blocks will timeout after 1 second if no blocks are received
-    const blocks = await this.l2BlockDownloader.getL2Blocks(1);
+    const blocks = await this.l2BlockDownloader.getBlocks(1);
     await this.handleL2Blocks(blocks);
     await this.commitCurrentL2BlockNumber();
   }
@@ -190,7 +191,7 @@ export class ServerWorldStateSynchronizer implements WorldStateSynchronizer {
         duration,
         isBlockOurs: result.isBlockOurs,
         ...l2Block.getStats(),
-      });
+      } satisfies L2BlockHandledStats);
     }
   }